Implement Google Analytics with a parallel route Summary: I was reading the docs on parallel routes and I think this is actually the proper way to do this.
Test Plan:
Google Analytics still works.
